+++ components/spider-storage/src/state.rs
@@ -1,11 +1,11 @@
 pub mod error;
 pub mod job_cache;
-pub mod server;
+pub mod runtime;
 pub mod service;
 
 pub use error::StorageServerError;
 pub use job_cache::JobCache;
-pub use server::{ServerRuntime, create_server_runtime};
+pub use runtime::{Runtime, create_server_runtime};
 pub use service::ServiceState;
 
 #[cfg(test)]
